Refactor ClaimRewards component
