+++ src/components/wallet/earn/Validate/ValidatorSuspended.vue
@@ -64,20 +64,6 @@ h4 {
     }
 }
 
-.disabled_input {
-    display: inline-block;
-    border-radius: var(--border-radius-sm);
-    color: gray;
-    background-color: var(--bg-light);
-    padding: 6px 14px;
-    white-space: nowrap;
-    width: 70%;
-}
-
-.disabled_input:focus-visible {
-    outline: 0;
-}
-
 .input_label {
     margin-bottom: 0.5rem;
 }
